avformat/avisynth: allow linking to libavisynth
Requires using AviSynth 3.7.0+, as that was when the ability to
even build libavisynth.a was added.

Dynamically loading AviSynth is still to be considered the proper
behavior, so the ability to link directly to the library must be
enabled explicitly by using -DAVS_NO_DYNAMIC_LOAD in the CFLAGS.

Why dynamic loading is still the default, even in this patch:

Due to the specifics of AviSynth's plugin interfaces and the
historical inertia of how most plugins were written, a statically
linked libavisynth on Windows will only work with most prebuilt
plugins if AviSynth+ was built with MSVC or Clang-CL.

The typical dynamic loading procedure of the demuxer sidesteps this
problem entirely, and can work with either normal MSVC builds of
the AviSynth+ core (and therefore 95%+ of existing plugins) or
with experimental MinGW-w64/GCC builds of the AviSynth+ core.

If MinGW-w64 was used to build AviSynth+, any and all C++ interface
plugins must also be rebuilt with MinGW-w64/[GCC|Clang] as well.

On non-Windows OSes, this is a non-issue; the standard compilers
are GCC or standard GCC-compatible Clang, and any plugins that
work on other OSes will have had to have been ported to use
GCC/Clang.