avisynth: allow linking to libavisynth
Really not sure about this. The dynamic loading sections have been
commented out behind #if 0s, and any place AVSC_API functions are
used, they had to be changed from avs_library.[function] to simply
[function].  I would prefer there to be an easy way for both to
coexist (like defining the 'avs_library.' string - including the
period - to nothing).

Requires using AviSynth 3.7.0+, as that was when the ability to
even build libavisynth.a was added.