avformat/avisynth: allow linking to libavisynth
Some systems have issues with dlopen and atexit, leading to
AviSynth simply not working even if it was built correctly.
Linux distributions that use musl instead of glibc are likely
the largest group affected by this, and hopefully this will
remediate those types of situations.

Requires using AviSynth 3.7.0+, as that was when the ability to
even build libavisynth.a was added.

Dynamically loading AviSynth is still to be considered the proper
behavior, so the ability to link directly to the library must be
enabled explicitly by using -DAVS_NO_DYNAMIC_LOAD in the CFLAGS.